Changed cell attributes. Added expansive BFS distance computation

+++ Height based flow maps/src/Cell.java
@@ -3,8 +3,8 @@ public class Cell {
     //String test;
     float height;
     int flow_direction;
-    float cell_x;
-    float cell_y;
+    int cell_x;
+    int cell_y;
     String title;
 
 }
